Patch from Jason to fix pywin32 bug 1392316.
Added support for non-English versions of Windows.  This required a
restructure of the way time zones are retrieved.  In particular, the
"StandardName" key for the configured time zone is in the regional
language, whereas all of the TimeZones keys are in English.  I re-wrote
GetIndexedTimeZones to make it more flexible so time zones can be
retrieved by the "Standard" name.

This internationalization support required the use of _winreg over
win32api.Reg*, due to the fact that _winreg returns Unicode results.

Note that 2 of 27 tests will fail on a non-English platform due to the
regional translation of the time zone's display name.  These failures do
not affect the functionality of the module.